refactor(core/ethereum): rename address.py to helpers.py
