+++ core/src/apps/ethereum/verify_message.py
@@ -6,7 +6,7 @@ from trezor.ui.layouts import confirm_signverify
 
 from apps.common.signverify import decode_message
 
-from .address import address_from_bytes, bytes_from_address
+from .helpers import address_from_bytes, bytes_from_address
 from .sign_message import message_digest
 
 if False:
