+++ core/tests/test_apps.ethereum.helpers.py
@@ -2,7 +2,7 @@ from common import *
 from apps.common.paths import HARDENED
 
 if not utils.BITCOIN_ONLY:
-    from apps.ethereum.address import address_from_bytes
+    from apps.ethereum.helpers import address_from_bytes
     from apps.ethereum.networks import NetworkInfo
